// src/modules/catalog/enums/service.enums.ts
export enum ServiceModality {
  PRESENCIAL = 'presencial',
  REMOTO = 'remoto',
  MIXTO = 'mixto',
}

export enum ServiceStatus {
  BORRADOR = 'borrador',
  PUBLICADO = 'publicado',
  PAUSADO = 'pausado',
  ARCHIVADO = 'archivado',
}

export enum ServicePriceType {
  FIJO = 'fijo',
  REFERENCIAL = 'referencial',
  COTIZABLE = 'cotizable',
}

export enum ServiceRequirementType {
  CLIENTE = 'cliente',
  DOCUMENTO = 'documento',
}